backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
better-exceptions>=0.3.0
pyjwt==2.9.0
bcrypt==4.2.0
python-multipart==0.0.20
sqlmodel==0.0.22
asyncpg==0.30.0
python-dotenv==1.0.1
email-validator==2.2.0
pydantic[email]==2.9.2
# better-auth==0.0.1  # Placeholder - actual Better Auth Python equivalent (commented out as no direct Python equivalent)
pytest==8.3.3
pytest-asyncio==0.23.8
httpx==0.27.2